MWPW-195600 | AI SEO Annotation Mode & Stream Infrastructure Cleanup #169
## Summary

This PR introduces the `aiSeoAnnotation` operation — a new tenant-side annotation mode that
runs directly on the host page — and removes the Milo `getConfig()` async dependency across
most modules in favor of reading from the already-available `window.streamConfig`.

---

## New: `aiSeoAnnotation` operation (`streamlibs/operations/aiSeoAnnotation/`)

- New operation mode that runs annotation directly on a tenant page without an iframe wrapper
- **AI text regeneration**: floating sparkle button appears on hover over any `<p>`, heading,
  `<li>`, etc. — clicking calls `/api/content-regeneration` with block context and `thread_id`,
  replacing the element text in-place
- **AI image regeneration**: floating button appears on hover over `<img>` elements — clicking
  opens a prompt overlay that calls `/api/image-generation` and swaps the image
- Both flows record changes as annotation edits so they persist through the normal
  Save/Push-to-DA flow

---

## Deleted: `streamlibs/standalone/html-review.js`

- The 837-line standalone bootstrap has been removed. Its responsibilities are now covered by
  `previewer.js`'s new `selfRender` IIFE (activated via `?daRenderingApp=stream`) plus the
  extracted `ai-seo-annotation.js` module
- Regen logic that was duplicated in that file is now the single source of truth in
  `aiSeoAnnotation/`

---

## Config / architecture refactoring

- **`CONFIG` object moved** from `scripts.js` to `streamlibs/utils/constants.js` — making it
  importable by `previewer.js` and other modules without circular deps
- **`getMapperEnv()` moved** from `scripts.js` to `streamlibs/utils/utils.js` and extended to
  read `mapperOrigin` from URL params, so it resolves correctly when running in a tenant context
- **`getConfig()` async calls removed** from `preflight.js`, `target/da.js`, `utils/utils.js`
  — all now read/write `window.streamConfig.streamMapper` directly, which is pre-populated in
  `initPreviewer` before any operation starts
- `initializeTokens` and `ensureStreamMapperForStandalone` are now synchronous

---

## Preflight in tenant context

- `preflight.js` no longer assumes `#preflight-operation-container` exists in the DOM — if not
  found, `addPfContainer()` creates it, allowing preflight to run inside a tenant page

---

## `previewer.js` enhancements

- Added `aiSeoAnnotation` case in `initiatePreviewer` — calls `annotationOperationOnHostPage()`
  \+ `attachRegenHandlers()`
- `isAnnotationOp()` helper covers both `annotation` and `aiSeoAnnotation` throughout
  `persist()`, `saveChanges()`, `refreshAnnotationCanvas()`, and the collab snapshot handler
- **Self-render IIFE**: when `?daRenderingApp=stream` is in the URL, `previewer.js` bootstraps
  itself (loads CSS from `mapperOrigin` + calls `initPreviewer`)
- `draftLocation` from collab snapshot and initial `previewParams` is stored in
  `window.streamConfig` and used as the primary DA content fetch URL

---

## Annotation CSS extracted

- `streamlibs/operations/annotation/annotation.css` now contains styles previously inlined
- `styles/styles.css` reduced by ~3,400 lines

---
