+++ streamlibs/operations/aiSeoAnnotation/ai-seo-annotation.css
@@ -1,0 +1,145 @@
+/* ── Text regeneration button ─────────────────────────────────────────────── */
+
+.stream-regen-btn {
+  position: fixed;
+  display: none;
+  align-items: center;
+  justify-content: center;
+  width: 28px;
+  height: 28px;
+  border-radius: 50%;
+  background: #1473e6;
+  border: none;
+  cursor: pointer;
+  z-index: 2147483500;
+  box-shadow: 0 2px 6px rgb(0 0 0 / 30%);
+  padding: 0;
+  transition: transform 0.1s, background 0.15s;
+}
+
+.stream-regen-btn:hover {
+  transform: scale(1.12);
+  background: #0d66d0;
+}
+
+.stream-regen-btn.stream-regen-visible { display: flex; }
+
+.stream-regen-btn.stream-regen-loading {
+  background: #888;
+  cursor: wait;
+  pointer-events: none;
+}
+
+.stream-regen-btn svg {
+  width: 15px;
+  height: 15px;
+  fill: #fff;
+}
+
+/* ── Image regeneration button ────────────────────────────────────────────── */
+
+.stream-img-regen-btn {
+  position: fixed;
+  display: none;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+  background: #1473e6;
+  border: none;
+  cursor: pointer;
+  z-index: 2147483500;
+  box-shadow: 0 2px 8px rgb(0 0 0 / 35%);
+  padding: 0;
+  transition: transform 0.1s, background 0.15s;
+}
+
+.stream-img-regen-btn:hover {
+  transform: scale(1.1);
+  background: #0d66d0;
+}
+
+.stream-img-regen-btn.stream-img-regen-visible { display: flex; }
+
+.stream-img-regen-btn svg {
+  width: 16px;
+  height: 16px;
+  fill: #fff;
+}
+
+/* ── Image prompt overlay ─────────────────────────────────────────────────── */
+
+.stream-img-prompt-overlay {
+  position: fixed;
+  background: #fff;
+  border-radius: 10px;
+  box-shadow: 0 6px 24px rgb(0 0 0 / 22%);
+  padding: 14px 16px;
+  z-index: 2147483600;
+  display: none;
+  flex-direction: column;
+  gap: 10px;
+  width: 300px;
+  font-family: 'Adobe Clean', 'Trebuchet MS', sans-serif;
+}
+
+.stream-img-prompt-overlay.stream-img-prompt-visible { display: flex; }
+
+.stream-img-prompt-overlay label {
+  font-size: 13px;
+  font-weight: 600;
+  color: #222;
+  margin: 0;
+}
+
+.stream-img-prompt-input {
+  border: 1px solid #ccc;
+  border-radius: 6px;
+  padding: 8px 10px;
+  font-size: 13px;
+  resize: vertical;
+  min-height: 64px;
+  outline: none;
+  font-family: inherit;
+  color: #222;
+  line-height: 1.4;
+}
+
+.stream-img-prompt-input:focus { border-color: #1473e6; }
+
+.stream-img-prompt-actions {
+  display: flex;
+  gap: 8px;
+  justify-content: flex-end;
+}
+
+.stream-img-prompt-cancel {
+  padding: 6px 14px;
+  border-radius: 6px;
+  border: 1px solid #ccc;
+  background: #fff;
+  cursor: pointer;
+  font-size: 13px;
+  color: #555;
+}
+
+.stream-img-prompt-cancel:hover { background: #f4f4f4; }
+
+.stream-img-prompt-submit {
+  padding: 6px 14px;
+  border-radius: 6px;
+  border: none;
+  background: #1473e6;
+  color: #fff;
+  cursor: pointer;
+  font-size: 13px;
+  font-weight: 600;
+}
+
+.stream-img-prompt-submit:disabled {
+  background: #888;
+  cursor: wait;
+}
+
+.stream-img-prompt-submit:hover:not(:disabled) { background: #0d66d0; }
