+++ streamlibs/operations/annotation.js
@@ -4,7 +4,7 @@
 import { fetchFigmaContent } from '../sources/figma.js';
 import { fetchDAContent } from '../sources/da.js';
 import { hydrateFragmentLinksInDaBlocks } from './edit/fragment-hydrate.js';
-import { miloLoadArea, getConfig } from '../utils/utils.js';
+import { miloLoadArea } from '../utils/utils.js';
 import { getDACompatibleHtml, postData } from '../target/da.js';
 import { createAnnotationState, createAnnotationUI } from './annotation/state.js';
 import { createAnnotationStore } from './annotation/store.js';
@@ -15,11 +15,10 @@ import createAssetServiceClient from './annotation/asset-service.js';
 import createAssetsPanelController from './annotation/assets-panel.js';
 import requestParentCollabRefresh from './annotation/collab-sync.js';
 
+// ── Module singletons ────────────────────────────────────────────────────────
+
 const annotationState = createAnnotationState();
 const annotationUI = createAnnotationUI();
-let cachedCleanHtml = '';
-const regenReplacements = [];
-let cachedPageMetadataHtml = null;
 const store = createAnnotationStore({ annotationState, annotationUI });
 const annotationService = createAnnotationServiceClient();
 const assetService = createAssetServiceClient();
@@ -54,6 +53,12 @@ assetsPanel.setOnAssetsChanged(() => {
   commentsPanel.renderCommentsPanel();
 });
 
+let cachedCleanHtml = '';
+let cachedPageMetadataHtml = null;
+const regenReplacements = [];
+
+// ── Preview DOM helpers (annotationOperation only) ───────────────────────────
+
 function normalizeDAImages(root) {
   root.querySelectorAll('img').forEach((img) => {
     if (img.src.includes('content.da.live') && img.parentElement.tagName !== 'PICTURE') {
@@ -71,7 +76,8 @@ async function getDADom() {
     return html;
   }
   if (source === 'da') {
-    const html = await fetchDAContent(window.streamConfig.contentUrl);
+    const cfg = window.streamConfig;
+    const html = await fetchDAContent(cfg.draftLocation || cfg.contentUrl);
     normalizeDAImages(html);
     return html;
   }
@@ -79,9 +85,7 @@ async function getDADom() {
 }
 
 async function initializePreview() {
-  document.body.querySelectorAll(':scope > header, :scope > main').forEach((element) => {
-    element.remove();
-  });
+  document.body.querySelectorAll(':scope > header, :scope > main').forEach((el) => el.remove());
   const htmlDom = await getDADom();
   const headerEle = document.createElement('header');
   const mainEle = document.createElement('main');
@@ -90,133 +94,19 @@ async function initializePreview() {
   if (cachedPageMetadataHtml !== null) {
     metadataEle.innerHTML = cachedPageMetadataHtml;
   } else {
-    const metadataBlocks = htmlDom.querySelectorAll('div.metadata');
-    if (metadataBlocks) {
-      metadataBlocks.forEach((mb) => {
-        metadataEle.innerHTML += mb.innerHTML;
-      });
-    }
+    htmlDom.querySelectorAll('div.metadata').forEach((mb) => {
+      metadataEle.innerHTML += mb.innerHTML;
+    });
   }
-  if (htmlDom instanceof HTMLElement && htmlDom.tagName === 'MAIN') {
-    mainEle.innerHTML = htmlDom.innerHTML;
-  } else {
-    mainEle.innerHTML = htmlDom;
-  }
+  mainEle.innerHTML = (htmlDom instanceof HTMLElement && htmlDom.tagName === 'MAIN')
+    ? htmlDom.innerHTML
+    : htmlDom;
   document.body.append(metadataEle);
   document.body.prepend(mainEle);
   document.body.prepend(headerEle);
 }
 
-function findAssetElement(doc, elementPath, elementProps, originalSrc) {
-  const main = doc.querySelector('main');
-  if (!main) return null;
-
-  if (elementPath) {
-    const selector = elementPath.startsWith('main > ')
-      ? elementPath.slice('main > '.length) : elementPath;
-    try {
-      const el = main.querySelector(selector);
-      if (el) return el;
-    } catch { /* invalid selector */ }
-  }
-
-  if (elementProps) {
-    const sectionIndex = typeof elementProps.sectionIndex === 'number' ? elementProps.sectionIndex : -1;
-    const blockClass = typeof elementProps.blockClass === 'string' ? elementProps.blockClass : '';
-    const blockIndex = typeof elementProps.blockIndex === 'number' ? elementProps.blockIndex : 0;
-
-    if (sectionIndex >= 0) {
-      const sections = Array.from(main.children).filter((el) => el.tagName === 'DIV');
-      const section = sections[sectionIndex];
-      if (section) {
-        let block = null;
-        if (blockClass) {
-          const matching = Array.from(section.querySelectorAll(`:scope > div.${blockClass}`));
-          block = matching[blockIndex] ?? matching[0] ?? null;
-        }
-        if (!block) {
-          const divs = Array.from(section.children).filter((el) => el.tagName === 'DIV');
-          block = divs[blockIndex] ?? null;
-        }
-        if (block) {
-          const pic = block.querySelector('picture');
-          if (pic && pic.querySelector('img')) return pic;
-          const img = block.querySelector('img');
-          if (img) return img;
-        }
-      }
-    }
-  }
-
-  if (originalSrc) {
-    const allImages = main.querySelectorAll('img');
-    const withoutParams = originalSrc.split('?')[0]?.split('#')[0] ?? '';
-    const filename = withoutParams.split('/').pop() || '';
-
-    const candidates = [];
-    for (const img of allImages) {
-      const src = img.getAttribute('src') || '';
-      if (src && withoutParams && src.includes(withoutParams)) {
-        candidates.push(img);
-      }
-    }
-    if (candidates.length === 0 && filename) {
-      for (const img of allImages) {
-        const src = (img.getAttribute('src') || '').split('?')[0]?.split('#')[0] ?? '';
-        if (src.split('/').pop() === filename) {
-          candidates.push(img);
-        }
-      }
-    }
-
-    if (candidates.length === 1) {
-      return candidates[0].closest('picture') || candidates[0];
-    }
-
-    if (candidates.length > 1 && elementProps) {
-      const sectionIndex = typeof elementProps.sectionIndex === 'number' ? elementProps.sectionIndex : -1;
-      const blockClass = typeof elementProps.blockClass === 'string' ? elementProps.blockClass : '';
-      const blockIndex = typeof elementProps.blockIndex === 'number' ? elementProps.blockIndex : 0;
-
-      if (sectionIndex >= 0) {
-        const sections = Array.from(main.children).filter((el) => el.tagName === 'DIV');
-        const section = sections[sectionIndex];
-        if (section) {
-          let block = null;
-          if (blockClass) {
-            const matching = Array.from(section.querySelectorAll(`:scope > div.${blockClass}`));
-            block = matching[blockIndex] ?? matching[0] ?? null;
-          }
-          if (!block) {
-            const divs = Array.from(section.children).filter((el) => el.tagName === 'DIV');
-            block = divs[blockIndex] ?? null;
-          }
-          if (block) {
-            const blockCandidate = candidates.find((img) => block.contains(img));
-            if (blockCandidate) return blockCandidate.closest('picture') || blockCandidate;
-          }
-        }
-      }
-    }
-
-    if (candidates.length > 0) {
-      return candidates[0].closest('picture') || candidates[0];
-    }
-  }
-
-  return null;
-}
-
-function replaceAssetUrl(element, assetUrl) {
-  const img = element.tagName === 'IMG' ? element : element.querySelector('img');
-  if (img) {
-    img.setAttribute('src', assetUrl);
-    if (img.hasAttribute('srcset')) img.setAttribute('srcset', assetUrl);
-  }
-  element.querySelectorAll('source').forEach((source) => {
-    source.setAttribute('srcset', assetUrl);
-  });
-}
+// ── URL / HTML helpers ────────────────────────────────────────────────────────
 
 function extractFilename(url) {
   if (!url) return '';
@@ -237,10 +127,7 @@ function toDaMediaUrl(url) {
 function rewriteAttr(el, attr, origin) {
   const val = el.getAttribute(attr);
   if (!val) return;
-  if (val.startsWith('data:')) {
-    el.setAttribute('data-regen-src', val);
-    return;
-  }
+  if (val.startsWith('data:')) { el.setAttribute('data-regen-src', val); return; }
   if (!origin) return;
   if (val.startsWith('./media')) {
     el.setAttribute(attr, `${origin}/${val.slice(2)}`);
@@ -260,10 +147,103 @@ function rewriteMediaUrls(container) {
   });
 }
 
+// ── Asset element resolution ──────────────────────────────────────────────────
+
+function replaceAssetUrl(element, assetUrl) {
+  const img = element.tagName === 'IMG' ? element : element.querySelector('img');
+  if (img) {
+    img.setAttribute('src', assetUrl);
+    if (img.hasAttribute('srcset')) img.setAttribute('srcset', assetUrl);
+  }
+  element.querySelectorAll('source').forEach((source) => {
+    source.setAttribute('srcset', assetUrl);
+  });
+}
+
+function resolveBlockInSection(section, blockClass, blockIndex) {
+  if (blockClass) {
+    const matching = Array.from(section.querySelectorAll(`:scope > div.${blockClass}`));
+    const block = matching[blockIndex] ?? matching[0] ?? null;
+    if (block) return block;
+  }
+  const divs = Array.from(section.children).filter((el) => el.tagName === 'DIV');
+  return divs[blockIndex] ?? null;
+}
+
+function findAssetElement(doc, elementPath, elementProps, originalSrc) {
+  const main = doc.querySelector('main');
+  if (!main) return null;
+
+  if (elementPath) {
+    const selector = elementPath.startsWith('main > ')
+      ? elementPath.slice('main > '.length) : elementPath;
+    try {
+      const el = main.querySelector(selector);
+      if (el) return el;
+    } catch { /* invalid selector */ }
+  }
+
+  const sectionIndex = typeof elementProps?.sectionIndex === 'number'
+    ? elementProps.sectionIndex : -1;
+  const blockClass = typeof elementProps?.blockClass === 'string' ? elementProps.blockClass : '';
+  const blockIndex = typeof elementProps?.blockIndex === 'number' ? elementProps.blockIndex : 0;
+
+  if (elementProps && sectionIndex >= 0) {
+    const sections = Array.from(main.children).filter((el) => el.tagName === 'DIV');
+    const section = sections[sectionIndex];
+    if (section) {
+      const block = resolveBlockInSection(section, blockClass, blockIndex);
+      if (block) {
+        const pic = block.querySelector('picture');
+        if (pic && pic.querySelector('img')) return pic;
+        const img = block.querySelector('img');
+        if (img) return img;
+      }
+    }
+  }
+
+  if (originalSrc) {
+    const allImages = main.querySelectorAll('img');
+    const withoutParams = originalSrc.split('?')[0]?.split('#')[0] ?? '';
+    const filename = withoutParams.split('/').pop() || '';
+    const candidates = [];
+
+    for (const img of allImages) {
+      const src = img.getAttribute('src') || '';
+      if (src && withoutParams && src.includes(withoutParams)) candidates.push(img);
+    }
+    if (candidates.length === 0 && filename) {
+      for (const img of allImages) {
+        const src = (img.getAttribute('src') || '').split('?')[0]?.split('#')[0] ?? '';
+        if (src.split('/').pop() === filename) candidates.push(img);
+      }
+    }
+
+    if (candidates.length === 1) return candidates[0].closest('picture') || candidates[0];
+
+    if (candidates.length > 1 && sectionIndex >= 0) {
+      const sections = Array.from(main.children).filter((el) => el.tagName === 'DIV');
+      const section = sections[sectionIndex];
+      if (section) {
+        const block = resolveBlockInSection(section, blockClass, blockIndex);
+        if (block) {
+          const blockCandidate = candidates.find((img) => block.contains(img));
+          if (blockCandidate) return blockCandidate.closest('picture') || blockCandidate;
+        }
+      }
+    }
+
+    if (candidates.length > 0) return candidates[0].closest('picture') || candidates[0];
+  }
+
+  return null;
+}
+
+// ── HTML export ───────────────────────────────────────────────────────────────
+
 function buildHtmlWithEditsAndAssets(assetReplacements) {
   const easyEdits = annotationState.store.easyEdits || [];
   const html = store.applyEasyEditsToHtmlString(cachedCleanHtml, easyEdits);
-
   const container = document.createElement('div');
   container.innerHTML = `<main>${html}</main>`;
 
@@ -273,45 +253,36 @@ function buildHtmlWithEditsAndAssets(assetReplacements) {
     );
     if (element) {
       replaceAssetUrl(element, asset.targetUrl);
-      // eslint-disable-next-line no-continue
-      continue;
-    }
-
-    const filenameCandidates = [
-      extractFilename(asset.originalSrc),
-      extractFilename(asset.daUrl),
-    ].filter(Boolean);
-
-    let matched = false;
-    if (filenameCandidates.length > 0) {
-      const allImages = container.querySelectorAll('img');
-      for (const img of allImages) {
-        const imgFilename = extractFilename(img.getAttribute('src') || '');
-        if (filenameCandidates.includes(imgFilename)) {
-          img.setAttribute('src', asset.targetUrl);
-          if (img.hasAttribute('srcset')) img.setAttribute('srcset', asset.targetUrl);
-          matched = true;
-          break;
+    } else {
+      const filenameCandidates = [
+        extractFilename(asset.originalSrc),
+        extractFilename(asset.daUrl),
+      ].filter(Boolean);
+      let matched = false;
+      if (filenameCandidates.length > 0) {
+        for (const img of container.querySelectorAll('img')) {
+          if (filenameCandidates.includes(extractFilename(img.getAttribute('src') || ''))) {
+            img.setAttribute('src', asset.targetUrl);
+            if (img.hasAttribute('srcset')) img.setAttribute('srcset', asset.targetUrl);
+            matched = true;
+            break;
+          }
         }
       }
-    }
-
-    if (!matched && asset.daUrl) {
-      const allImages = container.querySelectorAll('img');
-      for (const img of allImages) {
-        const src = img.getAttribute('src') || '';
-        if (src && src === asset.daUrl) {
-          img.setAttribute('src', asset.targetUrl);
-          if (img.hasAttribute('srcset')) img.setAttribute('srcset', asset.targetUrl);
-          break;
+      if (!matched && asset.daUrl) {
+        for (const img of container.querySelectorAll('img')) {
+          if ((img.getAttribute('src') || '') === asset.daUrl) {
+            img.setAttribute('src', asset.targetUrl);
+            if (img.hasAttribute('srcset')) img.setAttribute('srcset', asset.targetUrl);
+            break;
+          }
         }
       }
     }
   }
 
-  // Apply image-src easyEdits — these drive regen replacements and take
-  // priority over the assetReplacements pass above because they carry the
-  // final promoted/uploaded URL and the reliable original src for matching.
+  // image-src easyEdits take priority — they carry the final promoted URL and
+  // reliable original src, so they overwrite the assetReplacements pass above.
   const imageSrcEdits = (annotationState.store.easyEdits || [])
     .filter((e) => e?.editType === 'image-src' && e.to);
   for (const edit of imageSrcEdits) {
@@ -321,8 +292,7 @@ function buildHtmlWithEditsAndAssets(assetReplacements) {
 
   for (const regen of regenReplacements) {
     const regenFilename = extractFilename(regen.originalSrc);
-    const allImages = container.querySelectorAll('img');
-    for (const img of allImages) {
+    for (const img of container.querySelectorAll('img')) {
       const src = (img.getAttribute('src') || '').split('?')[0]?.split('#')[0] ?? '';
       if (src === regen.originalSrc || (regenFilename && src.split('/').pop() === regenFilename)) {
         const relUrl = toDaMediaUrl(regen.targetUrl);
@@ -352,14 +322,10 @@ function buildHtmlWithEditsAndAssets(assetReplacements) {
     metadataDiv.className = 'metadata';
     metadataDiv.innerHTML = pageMetadataDom.innerHTML;
     metadataDiv.querySelectorAll('p').forEach((p) => {
-      const ptag = p;
-      [...ptag.attributes].forEach((attr) => {
-        ptag.removeAttribute(attr.name);
-      });
+      [...p.attributes].forEach((attr) => p.removeAttribute(attr.name));
     });
     metadataDiv.querySelectorAll('img').forEach((img) => {
-      const daSrc = img.getAttribute('data-stream-original-src');
-      img.setAttribute('src', daSrc);
+      img.setAttribute('src', img.getAttribute('data-stream-original-src'));
     });
     const divWrapper = document.createElement('div');
     divWrapper.append(metadataDiv);
@@ -369,13 +335,28 @@ function buildHtmlWithEditsAndAssets(assetReplacements) {
   return { easyEdits, daCompatibleHtml: getDACompatibleHtml(mainEl.innerHTML) };
 }
 
+// ── Session lifecycle ─────────────────────────────────────────────────────────
+
+function prepareAnnotationSession({ preserveRemoteEditState = false } = {}) {
+  const previousAnnotationMode = annotationUI.annotationMode || 'comments';
+  const shouldRestoreInlineMode = annotationUI.inlineMode
+    && window.streamConfig?.inlineEditingAllowed !== false;
+  inlineEditing.resetInlineEditModeState();
+  annotationUI.annotationMode = shouldRestoreInlineMode ? 'edit' : previousAnnotationMode;
+  document.body.classList.add('annotation-mode');
+  if (!preserveRemoteEditState) {
+    annotationState.latestSavedEditsUpdatedAt = null;
+    annotationState.pendingRemoteEditsSnapshot = null;
+    annotationState.hasLoadedInitialEditsSnapshot = false;
+  }
+  return { shouldRestoreInlineMode };
+}
+
 async function finishAnnotationSession(mainEl, {
   preserveRemoteEditState,
   shouldRestoreInlineMode,
 }) {
-  await commentsPanel.setupAnnotationUI(mainEl, {
-    preserveRemoteEditState,
-  });
+  await commentsPanel.setupAnnotationUI(mainEl, { preserveRemoteEditState });
   if (annotationState.latestRemoteCollabSnapshot) {
     commentsPanel.applyRemoteCollabSnapshot(annotationState.latestRemoteCollabSnapshot, {
       includeEdits: false,
@@ -396,22 +377,96 @@ async function finishAnnotationSession(mainEl, {
   }
 }
 
-export async function annotationOperation(options = {}) {
-  const {
-    preserveRemoteEditState = false,
-  } = options;
-  const previousAnnotationMode = annotationUI.annotationMode || 'comments';
-  const shouldRestoreInlineMode = annotationUI.inlineMode
-    && window.streamConfig?.inlineEditingAllowed !== false;
+// ── Asset / edit processing (shared by persist and save) ──────────────────────
 
-  inlineEditing.resetInlineEditModeState();
-  annotationUI.annotationMode = shouldRestoreInlineMode ? 'edit' : previousAnnotationMode;
-  document.body.classList.add('annotation-mode');
-  if (!preserveRemoteEditState) {
-    annotationState.latestSavedEditsUpdatedAt = null;
-    annotationState.pendingRemoteEditsSnapshot = null;
-    annotationState.hasLoadedInitialEditsSnapshot = false;
+async function uploadAndDecideAssets() {
+  let newlyUploadedIds = [];
+  try {
+    newlyUploadedIds = await assetsPanel.uploadLocalAssets();
+  } catch (err) {
+    console.error('[annotation] Upload of local assets failed:', err);
   }
+  const appliedAssetIds = [...new Set([...assetsPanel.getAppliedAssetIds(), ...newlyUploadedIds])];
+  if (appliedAssetIds.length > 0) {
+    try {
+      await assetService.batchDecideAssets(appliedAssetIds, 'accepted');
+      assetsPanel.clearAppliedAssets();
+    } catch (err) {
+      console.error('[annotation] Batch decide failed:', err);
+    }
+  }
+  return newlyUploadedIds;
+}
+
+function buildAssetReplacementsAndEdits(resolveTargetUrl) {
+  const latestByPath = new Map();
+  for (const asset of (annotationState.store.assets || [])) {
+    // eslint-disable-next-line no-continue
+    if (!asset.originalSrc || !asset.daUrl) continue;
+    const existing = latestByPath.get(asset.elementPath);
+    const isNewer = existing && asset.createdAt
+      && new Date(asset.createdAt) > new Date(existing.createdAt || 0);
+    if (!existing || isNewer) {
+      latestByPath.set(asset.elementPath, asset);
+    }
+  }
+
+  const assetReplacements = Array.from(latestByPath.values()).map((asset) => ({
+    elementPath: asset.elementPath,
+    elementProps: asset.elementProps,
+    originalSrc: asset.originalSrc,
+    daUrl: asset.daUrl,
+    targetUrl: resolveTargetUrl(asset),
+  }));
+
+  for (const asset of latestByPath.values()) {
+    const finalUrl = resolveTargetUrl(asset);
+    if (!asset.elementPath || !finalUrl) continue; // eslint-disable-line no-continue
+    const existingEdit = store.getEasyEditByElement(
+      asset.elementRef || '', asset.elementPath, asset.elementProps,
+    );
+    if (!existingEdit || existingEdit.editType === 'image-src') {
+      store.upsertEasyEdit({
+        ...(existingEdit || {}),
+        editType: 'image-src',
+        elementPath: asset.elementPath,
+        elementProps: asset.elementProps || {},
+        elementRef: asset.elementRef || '',
+        from: existingEdit?.from || asset.originalSrc,
+        to: finalUrl,
+        fromHtml: '',
+        toHtml: '',
+        updatedAt: new Date().toISOString(),
+      });
+    }
+  }
+
+  return assetReplacements;
+}
+
+// ── da.live URL normalizer ────────────────────────────────────────────────────
+
+/** da.live edit URL → repo path (e.g. adobecom/...); leaves plain paths unchanged. */
+function normalizePersistUrlForDaApi(raw) {
+  const s = `${raw || ''}`.trim();
+  if (!s) return '';
+  if (!/^https?:\/\//i.test(s)) return s.replace(/^\/+/, '');
+  try {
+    const u = new URL(s);
+    if (u.hash?.startsWith('#/')) return decodeURIComponent(u.hash.slice(2)).replace(/^\/+/, '');
+    if (u.hostname.includes('da.live') && u.pathname && u.pathname !== '/') {
+      return decodeURIComponent(u.pathname).replace(/^\/+/, '');
+    }
+  } catch { /* ignore */ }
+  return s;
+}
+
+// ── Public API ────────────────────────────────────────────────────────────────
+
+export async function annotationOperation(options = {}) {
+  const { preserveRemoteEditState = false } = options;
+  const { shouldRestoreInlineMode } = prepareAnnotationSession({ preserveRemoteEditState });
+
   await initializePreview();
   const mainEl = document.querySelector('main');
   if (!mainEl) return;
@@ -422,9 +477,7 @@ export async function annotationOperation(options = {}) {
     });
   }
 
-  if (!cachedCleanHtml) {
-    cachedCleanHtml = mainEl.innerHTML || '';
-  }
+  if (!cachedCleanHtml) cachedCleanHtml = mainEl.innerHTML || '';
 
   if (window.streamConfig?.source === 'da') {
     const insertedFragments = await hydrateFragmentLinksInDaBlocks(mainEl);
@@ -436,20 +489,16 @@ export async function annotationOperation(options = {}) {
 
   await miloLoadArea();
 
-  // initialize page metadata
   const metadataDom = document.body.querySelector('.page-metadata');
   const metadataSeparator = document.createElement('div');
-  metadataSeparator.classList.add('section');
-  metadataSeparator.classList.add('stream-annotation-page-metadata');
+  metadataSeparator.classList.add('section', 'stream-annotation-page-metadata');
   metadataSeparator.innerHTML = '<h3>Page Metadata</h3>';
   metadataSeparator.append(metadataDom);
 
-  function addAndRegisterRow(row) {
+  const addAndRegisterRow = (row) => {
     metadataDom.append(row);
-    row.querySelectorAll('p').forEach((p) => {
-      inlineEditing.registerNewEditableElement(p);
-    });
-  }
+    row.querySelectorAll('p').forEach((p) => inlineEditing.registerNewEditableElement(p));
+  };
 
   const addTextBtn = document.createElement('button');
   addTextBtn.className = 'stream-annotation-add-metadata-row';
@@ -471,15 +520,11 @@ export async function annotationOperation(options = {}) {
 
   const metadataActions = document.createElement('div');
   metadataActions.className = 'stream-annotation-metadata-actions';
-  metadataActions.append(addTextBtn);
-  metadataActions.append(addImageBtn);
+  metadataActions.append(addTextBtn, addImageBtn);
   metadataSeparator.append(metadataActions);
   mainEl.append(metadataSeparator);
 
-  await finishAnnotationSession(mainEl, {
-    preserveRemoteEditState,
-    shouldRestoreInlineMode,
-  });
+  await finishAnnotationSession(mainEl, { preserveRemoteEditState, shouldRestoreInlineMode });
 }
 
 export async function annotationOperationOnHostPage(options = {}) {
@@ -488,28 +533,16 @@ export async function annotationOperationOnHostPage(options = {}) {
     refreshBaselineHtml = false,
     baselineHtml = null,
   } = options;
-  const previousAnnotationMode = annotationUI.annotationMode || 'comments';
-  const shouldRestoreInlineMode = annotationUI.inlineMode
-    && window.streamConfig?.inlineEditingAllowed !== false;
-
-  inlineEditing.resetInlineEditModeState();
-  annotationUI.annotationMode = shouldRestoreInlineMode ? 'edit' : previousAnnotationMode;
-  document.body.classList.add('annotation-mode');
-  if (!preserveRemoteEditState) {
-    annotationState.latestSavedEditsUpdatedAt = null;
-    annotationState.pendingRemoteEditsSnapshot = null;
-    annotationState.hasLoadedInitialEditsSnapshot = false;
-  }
+  const { shouldRestoreInlineMode } = prepareAnnotationSession({ preserveRemoteEditState });
 
   const mainEl = document.querySelector('main');
-  if (!mainEl) {
-    throw new Error('annotationOperationOnHostPage: no <main> found on page');
-  }
+  if (!mainEl) throw new Error('annotationOperationOnHostPage: no <main> found on page');
 
   if (!cachedCleanHtml || refreshBaselineHtml) {
     if (window.streamConfig?.source === 'da') {
       try {
-        const daMain = await fetchDAContent(window.streamConfig.contentUrl);
+        const cfg = window.streamConfig;
+        const daMain = await fetchDAContent(cfg.draftLocation || cfg.contentUrl);
         cachedCleanHtml = daMain?.innerHTML || '';
       } catch (err) {
         console.warn('[annotation] Failed to fetch DA baseline HTML, falling back to live DOM:', err);
@@ -520,53 +553,12 @@ export async function annotationOperationOnHostPage(options = {}) {
     }
   }
 
-  await finishAnnotationSession(mainEl, {
-    preserveRemoteEditState,
-    shouldRestoreInlineMode,
-  });
-}
-
-/** da.live edit URL → repo path (e.g. adobecom/...); leaves plain paths unchanged. */
-function normalizePersistUrlForDaApi(raw) {
-  const s = `${raw || ''}`.trim();
-  if (!s) return '';
-  if (!/^https?:\/\//i.test(s)) {
-    return s.replace(/^\/+/, '');
-  }
-  try {
-    const u = new URL(s);
-    if (u.hash?.startsWith('#/')) {
-      return decodeURIComponent(u.hash.slice(2)).replace(/^\/+/, '');
-    }
-    if (u.hostname.includes('da.live') && u.pathname && u.pathname !== '/') {
-      return decodeURIComponent(u.pathname).replace(/^\/+/, '');
-    }
-  } catch {
-    /* ignore */
-  }
-  return s;
+  await finishAnnotationSession(mainEl, { preserveRemoteEditState, shouldRestoreInlineMode });
 }
 
 export async function persistAnnotationChangesToDA() {
   await inlineEditing.syncInlineEditsBeforePersist();
-
-  let newlyUploadedIds = [];
-  try {
-    newlyUploadedIds = await assetsPanel.uploadLocalAssets();
-  } catch (err) {
-    console.error('[annotation] Upload of local assets failed before promote:', err);
-  }
-
-  const fromApplied = assetsPanel.getAppliedAssetIds();
-  const appliedAssetIds = [...new Set([...fromApplied, ...newlyUploadedIds])];
-  if (appliedAssetIds.length > 0) {
-    try {
-      await assetService.batchDecideAssets(appliedAssetIds, 'accepted');
-      assetsPanel.clearAppliedAssets();
-    } catch (err) {
-      console.error('[annotation] Batch decide failed:', err);
-    }
-  }
+  await uploadAndDecideAssets();
 
   let promotedAssets = [];
   try {
@@ -582,48 +574,9 @@ export async function persistAnnotationChangesToDA() {
     else annotationState.store.assets.push(promoted);
   }
 
-  const latestByPath = new Map();
-  for (const asset of (annotationState.store.assets || [])) {
-    // eslint-disable-next-line no-continue
-    if (!asset.originalSrc || !asset.daUrl) continue;
-    const existing = latestByPath.get(asset.elementPath);
-    // eslint-disable-next-line max-len
-    if (!existing || (asset.createdAt && new Date(asset.createdAt) > new Date(existing.createdAt || 0))) {
-      latestByPath.set(asset.elementPath, asset);
-    }
-  }
-
-  const assetReplacements = Array.from(latestByPath.values()).map((asset) => ({
-    elementPath: asset.elementPath,
-    elementProps: asset.elementProps,
-    originalSrc: asset.originalSrc,
-    daUrl: asset.daUrl,
-    targetUrl: asset.finalDaUrl || asset.daUrl,
-  }));
-
-  for (const asset of latestByPath.values()) {
-    const finalUrl = asset.finalDaUrl || asset.daUrl;
-    // eslint-disable-next-line no-continue
-    if (!asset.elementPath || !finalUrl) continue;
-    const existingEdit = store.getEasyEditByElement(
-      asset.elementRef || '', asset.elementPath, asset.elementProps,
-    );
-    if (!existingEdit || existingEdit.editType === 'image-src') {
-      store.upsertEasyEdit({
-        ...(existingEdit || {}),
-        editType: 'image-src',
-        elementPath: asset.elementPath,
-        elementProps: asset.elementProps || {},
-        elementRef: asset.elementRef || '',
-        from: existingEdit?.from || asset.originalSrc,
-        to: finalUrl,
-        fromHtml: '',
-        toHtml: '',
-        updatedAt: new Date().toISOString(),
-      });
-    }
-  }
-
+  const assetReplacements = buildAssetReplacementsAndEdits(
+    (asset) => asset.finalDaUrl || asset.daUrl,
+  );
   const { daCompatibleHtml } = buildHtmlWithEditsAndAssets(assetReplacements);
 
   const cfg = window.streamConfig || {};
@@ -633,80 +586,19 @@ export async function persistAnnotationChangesToDA() {
       'persistAnnotationChangesToDA: streamConfig.pageUrl is required (set via STREAM_HTML_REVIEW_INIT).',
     );
   }
-  const pushUrl = normalizePersistUrlForDaApi(rawPushUrl) || rawPushUrl;
-
-  await postData(pushUrl, daCompatibleHtml, {
+  await postData(normalizePersistUrlForDaApi(rawPushUrl) || rawPushUrl, daCompatibleHtml, {
     suppressErrorPage: true,
   });
 }
 
 export async function saveAnnotationChanges(reportProgress = () => {}) {
   await inlineEditing.syncInlineEditsBeforePersist();
+  await uploadAndDecideAssets();
 
-  let newlyUploadedIds = [];
-  try {
-    newlyUploadedIds = await assetsPanel.uploadLocalAssets();
-  } catch (err) {
-    console.error('[annotation] Bulk upload of local assets failed:', err);
-  }
-
-  const fromApplied = assetsPanel.getAppliedAssetIds();
-  const appliedAssetIds = [...new Set([...fromApplied, ...newlyUploadedIds])];
-  if (appliedAssetIds.length > 0) {
-    try {
-      await assetService.batchDecideAssets(appliedAssetIds, 'accepted');
-      assetsPanel.clearAppliedAssets();
-    } catch (err) {
-      console.error('[annotation] Batch decide failed:', err);
-    }
-  }
-
-  const latestByPath = new Map();
-  for (const asset of (annotationState.store.assets || [])) {
-    // eslint-disable-next-line no-continue
-    if (!asset.originalSrc || !asset.daUrl) continue;
-    const existing = latestByPath.get(asset.elementPath);
-    // eslint-disable-next-line max-len
-    if (!existing || (asset.createdAt && new Date(asset.createdAt) > new Date(existing.createdAt || 0))) {
-      latestByPath.set(asset.elementPath, asset);
-    }
-  }
-  const assetReplacements = Array.from(latestByPath.values()).map((asset) => ({
-    elementPath: asset.elementPath,
-    elementProps: asset.elementProps,
-    originalSrc: asset.originalSrc,
-    daUrl: asset.daUrl,
-    targetUrl: asset.daUrl,
-  }));
-
-  // Record each uploaded regen asset as an image-src easyEdit so the
-  // change is tracked in history and drives reliable DOM replacement.
-  for (const asset of latestByPath.values()) {
-    if (!asset.elementPath || !asset.daUrl) continue; // eslint-disable-line no-continue
-    const existingEdit = store.getEasyEditByElement(
-      asset.elementRef || '', asset.elementPath, asset.elementProps,
-    );
-    if (!existingEdit || existingEdit.editType === 'image-src') {
-      store.upsertEasyEdit({
-        ...(existingEdit || {}),
-        editType: 'image-src',
-        elementPath: asset.elementPath,
-        elementProps: asset.elementProps || {},
-        elementRef: asset.elementRef || '',
-        from: existingEdit?.from || asset.originalSrc,
-        to: asset.daUrl,
-        fromHtml: '',
-        toHtml: '',
-        updatedAt: new Date().toISOString(),
-      });
-    }
-  }
-
+  const assetReplacements = buildAssetReplacementsAndEdits((asset) => asset.daUrl);
   const { easyEdits, daCompatibleHtml } = buildHtmlWithEditsAndAssets(assetReplacements);
 
-  await postData(window.streamConfig.targetUrl, daCompatibleHtml, {
-    suppressErrorPage: true,
-  });
+  await postData(window.streamConfig.targetUrl, daCompatibleHtml, { suppressErrorPage: true });
   reportProgress('htmlSaved');
 
   if (annotationService.isAvailable()) {
@@ -740,13 +632,11 @@ export function recordTextRegenAsEdit(element, fromText, toText, fromHtml = '') 
 
   const editAnchor = store.buildEditElementAnchor(element, annotationUI.mainEl);
   const segments = store.getChangedSegments(baselineText, toText);
-
   const existing = store.getEasyEditByElement(
-    elementRef,
-    editAnchor.elementPath,
-    editAnchor.elementProps,
+    elementRef, editAnchor.elementPath, editAnchor.elementProps,
   );
-  const editRecord = {
+
+  const persistedEdit = store.upsertEasyEdit({
     id: existing?.id || store.generateId('easy-edit'),
     editType: 'text',
     attrName: '',
@@ -760,13 +650,10 @@ export function recordTextRegenAsEdit(element, fromText, toText, fromHtml = '') 
     changedFrom: segments.changedFrom,
     changedTo: segments.changedTo,
     updatedAt: new Date().toISOString(),
-  };
-
-  const persistedEdit = store.upsertEasyEdit(editRecord);
+  });
 
   const editThread = store.getEditThreadByElementPath(
-    persistedEdit?.elementPath,
-    persistedEdit?.elementProps,
+    persistedEdit?.elementPath, persistedEdit?.elementProps,
   );
   if (editThread) {
     annotationState.activeThreadId = editThread.id;
@@ -784,12 +671,10 @@ export function registerRegenReplacement(originalSrc, newUrl) {
   else regenReplacements.push({ originalSrc, targetUrl: newUrl });
 }
 
-/* eslint-disable no-console */
 export async function recordImageRegenAsLocalAsset(imgEl, generatedUrl) {
   if (!(imgEl instanceof HTMLImageElement) || !generatedUrl) return;
 
-  const cfg = await getConfig();
-  const rawToken = cfg?.streamMapper?.daToken || window.streamConfig?.token || '';
+  const rawToken = window.streamConfig?.streamMapper?.daToken || window.streamConfig?.token || '';
   const authToken = rawToken && !rawToken.startsWith('Bearer ') ? `Bearer ${rawToken}` : rawToken;
 
   let blob;
@@ -824,9 +709,7 @@ export function preparePendingRemoteEditsRefresh() {
 
 export async function refreshAnnotationFloatingUI() {
   await new Promise((resolve) => {
-    window.requestAnimationFrame(() => {
-      window.requestAnimationFrame(resolve);
-    });
+    window.requestAnimationFrame(() => window.requestAnimationFrame(resolve));
   });
   commentsPanel.renderThreadMarkers({ resolveTargets: true });
 }
