+++ streamlibs/operations/edit/edit.css
@@ -3,6 +3,7 @@
 body.editor-mode {
   height: 100vh;
   overflow: hidden;
+  background: #c8c9ca;
 }
 
 body.editor-mode .da-panel {
@@ -15,6 +16,7 @@ body.editor-mode .da-panel {
   overflow: auto;
   scale: 0.5;
   transform-origin: top right;
+  background: white;
 }
 
 body.editor-mode .figma-panel {
@@ -27,7 +29,7 @@ body.editor-mode .figma-panel {
   overflow: auto;
   scale: 0.5;
   transform-origin: top left;
-  background: #c8c9ca;
+  background: white;
   border-right: 3px solid #e7e0e0;
 }
 
@@ -120,6 +122,11 @@ div[data-source="figma"] {
   cursor: grab;
 }
 
+div[data-source="figma"][data-placeholder="true"],
+div[data-source="figma"].do-not-drag {
+  cursor: default;
+}
+
 .section[data-source="figma"],
 .section[data-source="da"] {
   border-top: 1px dashed #8a8a8a;
@@ -147,20 +154,6 @@ div[data-source="figma"] {
   right: 30px;
 }
 
-@keyframes alert {
-  0% {
-    box-shadow: 0 0 0 0 rgba(253, 94, 83, 70%);
-  }
-
-  70% {
-    box-shadow: 0 0 0 20px rgba(253, 94, 83, 0%);
-  }
-
-  100% {
-    box-shadow: 0 0 0 0 rgba(253, 94, 83, 0%);
-  }
-}
-
 .da-panel.da-drop-active {
   outline: 3px dashed var(--sm-color-primary);
   outline-offset: 4px;
@@ -169,6 +162,7 @@ div[data-source="figma"] {
 .da-panel > [data-source="figma"],
 .da-panel > [data-source="da"] {
   position: relative;
+  min-height: 100px;
 }
 
 .da-panel > [data-source="figma"]::before {
@@ -186,7 +180,7 @@ div[data-source="figma"] {
   transition: width 0.5s ease-in-out;
 }
 
-.da-panel > [data-source="figma"]:hover::before{
+.da-panel > [data-source="figma"]:hover::before {
   width: 20px;
 }
 
@@ -257,7 +251,8 @@ div.figma-panel.hidden {
   display: none;
 }
 
-.figma-panel .section:not(:has(div)),
-.da-panel .section:not(:has(div)) {
-    display: none;
+@keyframes alert {
+  0% { box-shadow: 0 0 0 0 rgba(253, 94, 83, 70%); }
+  70% { box-shadow: 0 0 0 20px rgba(253, 94, 83, 0%); }
+  100% { box-shadow: 0 0 0 0 rgba(253, 94, 83, 0%); }
 }
