+++ streamlibs/operations/preflight/preflight.css
@@ -1,4 +1,7 @@
 /* ============ PREFLIGHT OPERATION =========== */
+/* stylelint-disable color-function-notation */
+/* stylelint-disable color-function-alias-notation */
+/* stylelint-disable alpha-value-notation */
   
 .preflight-iframe {
   width: 100%;
