+++ streamlibs/scripts.js
@@ -1,4 +1,5 @@
-import { setLibs } from './utils/utils.js';
+import { setLibs, getMapperEnv } from './utils/utils.js';
+import { CONFIG } from './utils/constants.js';
 
 // eslint-disable-next-line no-unused-vars
 function decorateArea(area = document) {
@@ -20,85 +21,6 @@ function decorateArea(area = document) {
 const STYLES = '';
 const LIBS = '/libs';
 
-function getMapperEnv() {
-  const { origin } = window.location;
-  if (origin.includes('dev--')) return 'dev';
-  if (origin.includes('dev02--')) return 'dev02';
-  if (origin.includes('stage--')) return 'stage';
-  if (origin.includes('main--')) return 'prod';
-  return 'dev';
-}
-
-const CONFIG = {
-  decorateArea,
-  locales: {
-    '': { ietf: 'en-US', tk: 'hah7vzn.css' },
-    de: { ietf: 'de-DE', tk: 'hah7vzn.css' },
-    kr: { ietf: 'ko-KR', tk: 'zfo3ouc' },
-  },
-  figmaServiceRetry: {
-    retryCount: 3,
-    retryDelaysMs: [1000, 2000, 4000, 6000],
-    blockContentConcurrency: 3,
-  },
-  prod: {
-    streamMapper: {
-      serviceEP: 'https://adobe-acom-stream-service-deploy-ethos501-prod-or2-ab8ae6.cloud.adobe.io',
-      figmaMappingUrl: '/api/fig-comps',
-      figmaBlockContentUrl: '/api/fig-comp-details',
-      pushToDaUrl: '/api/push-html',
-      blockMappingsUrl: 'https://main--stream-mapper--adobecom.aem.live/block-mappings',
-      figmaAuthToken: '',
-      daToken: '',
-      preflightUrl: '/drafts/stream/tools/preflight-controller?milolibs=stream-prod',
-      sidekickLoginUrl: '/drafts/stream/tools/sidekick-controller?milolibs=stream-prod',
-      allowMessagesFromDomains: ['https://440859-stream.adobeio-static.net'],
-    },
-  },
-  stage: {
-    streamMapper: {
-      serviceEP: 'https://adobe-acom-stream-service-deploy-ethos502-prod-or2-32c93a.cloud.adobe.io',
-      figmaMappingUrl: '/api/fig-comps',
-      figmaBlockContentUrl: '/api/fig-comp-details',
-      pushToDaUrl: '/api/push-html',
-      blockMappingsUrl: 'https://stage--stream-mapper--adobecom.aem.page/block-mappings',
-      figmaAuthToken: '',
-      daToken: '',
-      preflightUrl: '/drafts/stream/tools/preflight-controller?milolibs=stream-stage',
-      sidekickLoginUrl: '/drafts/stream/tools/sidekick-controller?milolibs=stream-stage',
-      allowMessagesFromDomains: ['https://440859-stream*.adobeio-static.net'],
-    },
-  },
-  dev: {
-    streamMapper: {
-      serviceEP: 'https://adobe-acom-stream-service-deploy-ethos502-prod-or2-1de07c.cloud.adobe.io',
-      figmaMappingUrl: '/api/fig-comps',
-      figmaBlockContentUrl: '/api/fig-comp-details',
-      pushToDaUrl: '/api/push-html',
-      blockMappingsUrl: 'https://stage--stream-mapper--adobecom.aem.page/block-mappings',
-      figmaAuthToken: '',
-      daToken: '',
-      preflightUrl: '/drafts/stream/tools/preflight-controller?milolibs=stream-dev',
-      sidekickLoginUrl: '/drafts/stream/tools/sidekick-controller?milolibs=stream-dev',
-      allowMessagesFromDomains: ['*', 'https://440859-stream*.adobeio-static.net'],
-    },
-  },
-  dev02: {
-    streamMapper: {
-      serviceEP: 'https://adobe-acom-stream-service-deploy-ethos501-prod-or2-b0c6b7.cloud.adobe.io',
-      figmaMappingUrl: '/api/fig-comps',
-      figmaBlockContentUrl: '/api/fig-comp-details',
-      pushToDaUrl: '/api/push-html',
-      blockMappingsUrl: 'https://stage--stream-mapper--adobecom.aem.page/block-mappings',
-      figmaAuthToken: '',
-      daToken: '',
-      preflightUrl: '/drafts/stream/tools/preflight-controller?milolibs=stream-dev',
-      sidekickLoginUrl: '/drafts/stream/tools/sidekick-controller?milolibs=stream-dev',
-      allowMessagesFromDomains: ['*', 'https://440859-stream*.adobeio-static.net'],
-    },
-  },
-};
-
 decorateArea();
 
 const miloLibs = setLibs(LIBS);
@@ -117,7 +39,9 @@ const miloLibs = setLibs(LIBS);
 (async function loadPage() {
   const { loadArea, setConfig } = await import(`${miloLibs}/utils/utils.js`);
   // eslint-disable-next-line no-unused-vars
-  const config = setConfig({ ...CONFIG, ...CONFIG[getMapperEnv()], miloLibs });
+  const config = setConfig({
+    ...CONFIG, ...CONFIG[getMapperEnv()], decorateArea, miloLibs,
+  });
   await loadArea();
   const metaTag = document.querySelector('meta[name="initiate-previewer"]');
   if (metaTag && metaTag.getAttribute('content') === 'off') return;
